coliziuni si inamici

+++ core/src/main/java/com/acsurvivors/entities/components/TransformComponent.java
@@ -1,7 +1,7 @@
 package com.acsurvivors.entities.components;
 
 public class TransformComponent {
-    public int x, y;
+    public float x, y;
     public float scaleX = 1f, scaleY = 1f;
     public float rotation = 0f;
 }
